ISPN-9599 DefaultCacheManager.getGlobalComponentRegistry  should require ADMIN permission
